Changed Controller to Presenter and optimised the fragment transaction code
