+++ app/src/main/java/org/oppia/app/player/exploration/ExplorationFragment.kt
@@ -10,7 +10,7 @@ import javax.inject.Inject
 
 /** Fragment that contains displays single exploration. */
 class ExplorationFragment : InjectableFragment() {
-  @Inject lateinit var explorationFragmentController: ExplorationFragmentController
+  @Inject lateinit var explorationFragmentPresenter: ExplorationFragmentPresenter
 
   override fun onAttach(context: Context?) {
     super.onAttach(context)
@@ -18,6 +18,6 @@ class ExplorationFragment : InjectableFragment() {
   }
 
   override fun onCreateView(inflater: LayoutInflater, container: ViewGroup?, savedInstanceState: Bundle?): View? {
-    return explorationFragmentController.handleCreateView(inflater, container)
+    return explorationFragmentPresenter.handleCreateView(inflater, container)
   }
 }
