+++ app/src/main/java/org/oppia/app/player/exploration/ExplorationFragmentPresenter.kt
@@ -10,7 +10,7 @@ import javax.inject.Inject
 
 /** The controller for [ExplorationFragment]. */
 @FragmentScope
-class ExplorationFragmentController @Inject constructor(
+class ExplorationFragmentPresenter @Inject constructor(
   private val fragment: Fragment
 ) {
   fun handleCreateView(inflater: LayoutInflater, container: ViewGroup?): View? {
